refactoring
gotta keep em separated (your concerns i mean)
