+++ components/styles/PianoRoll.css
