+++ components/styles/Sequencer.css
