+++ components/styles/SequencerItem.css
