+++ components/styles/SequencerSelection.css
